Rename settingsMerger to nodePathHelper
